moved download and save functions into cohortDetails
