+++ src/pages/inventory/cohortModal/cohortModalGenerator.js
@@ -1,10 +1,5 @@
-import React, { useState, useContext, useEffect, useRef } from 'react';
+import React, { useState, useContext, useEffect } from 'react';
 import { CohortStateContext } from '../../../components/CohortSelectorState/CohortStateContext.js';
-import {
-    onDeleteSingleCohort,
-    onDeleteAllCohort,
-    onMutateSingleCohort,
-} from '../../../components/CohortSelectorState/store/action.js';
 import {
     Modal, withStyles,
 } from '@material-ui/core';
@@ -15,9 +10,8 @@ import CohortDetails from './components/cohortDetails';
 import DeleteConfirmationModal from './components/deleteConfirmationModal';
 import { deletionTypes } from './components/deleteConfirmationModal';
 import Alert from '@material-ui/lab/Alert';
-import { GET_COHORT_MANIFEST_QUERY, GET_COHORT_METADATA_QUERY } from '../../../bento/dashboardTabData.js';
-import client from '../../../utils/graphqlClient.js'
-import { arrayToCSVDownload, objectToJsonDownload, hasUnsavedChanges } from './utils.js';
+import { hasUnsavedChanges } from './utils.js';
+
 import { CohortModalContext } from './CohortModalContext.js'
 
 /**
@@ -33,7 +27,7 @@ export const CohortModalGenerator = (uiConfig = DEFAULT_CONFIG) => {
     } = uiConfig;
 
     const { currentCohortChanges, setCurrentCohortChanges } = useContext(CohortModalContext);
-    const { state, dispatch } = useContext(CohortStateContext);
+    const { state } = useContext(CohortStateContext);
     const [selectedCohort, setSelectedCohort] = useState(null); // Default to the first entry
     const [alert, setAlert] = useState({ type: '', message: '' });
     const ignoredFields = ["cohortId"]
@@ -62,23 +56,6 @@ export const CohortModalGenerator = (uiConfig = DEFAULT_CONFIG) => {
         ? config.title
         : DEFAULT_CONFIG.config.title;
 
-    const downloadCohortManifest = async () => {
-        const participantPKs = state[selectedCohort].participants.map(item => item.participant_pk);
-        const { data } = await client.query({
-            query: GET_COHORT_MANIFEST_QUERY,
-            variables: { "participant_pk": participantPKs, "first": state[selectedCohort].participants.length },
-        });
-        arrayToCSVDownload(data['diagnosisOverview'], selectedCohort);
-    };
-
-    const downloadCohortMetadata = async () => {
-        const participantPKs = state[selectedCohort].participants.map(item => item.participant_pk);
-        const { data } = await client.query({
-            query: GET_COHORT_METADATA_QUERY,
-            variables: { "participant_pk": participantPKs, "first": state[selectedCohort].participants.length },
-        });
-        objectToJsonDownload(data['cohortMetadata'], selectedCohort);
-    };
 
     // Handle saving updates to cohort
     const handleSetCurrentCohortChanges = (localCohort) => {
@@ -100,24 +77,6 @@ export const CohortModalGenerator = (uiConfig = DEFAULT_CONFIG) => {
         setCurrentCohortChanges(null);
     };
 
-    const handleSaveCohort = (localCohort) => {
-        if (!localCohort.cohortId) return;
-        dispatch(onMutateSingleCohort(
-            localCohort.cohortId,
-            {
-                cohortName: localCohort.cohortName,
-                cohortDescription: localCohort.cohortDescription,
-                participants: localCohort.participants
-            },
-            () => {
-                setAlert({ type: 'success', message: 'Cohort updated successfully!' });
-                handleClearCurrentCohortChanges();        
-            },
-            (error) => {
-                setAlert({ type: 'error', message: `Failed to update cohort: ${error.message}` });
-            } 
-        ));
-    };
 
     return {
         CohortModal: withStyles(DEFAULT_STYLES, { withTheme: true })((props) => {
@@ -202,11 +161,10 @@ export const CohortModalGenerator = (uiConfig = DEFAULT_CONFIG) => {
                                     activeCohort={state[selectedCohort]}
                                     temporaryCohort={currentCohortChanges}
                                     closeModal={unSavedChangesCheck}
-                                    handleSaveCohort={handleSaveCohort}
                                     handleSetCurrentCohortChanges={handleSetCurrentCohortChanges}
-                                    downloadCohortManifest={downloadCohortManifest}
-                                    downloadCohortMetadata={downloadCohortMetadata}
                                     deleteConfirmationClasses={deleteConfirmationClasses}
+                                    setAlert={setAlert}
+                                    handleClearCurrentCohortChanges={handleClearCurrentCohortChanges}
                                 />
                             </div>
                         </div>
